+++ lib/google-finance/api.rb
@@ -1,0 +1,2 @@
+require_relative 'api/index'
+require_relative 'api/get_prices'