+++ lib/google-finance/price.rb
@@ -1,0 +1,11 @@
+module GoogleFinance
+  class Price < Resource
+    property 'date'
+    property 'close'
+    property 'symbol'
+    property 'high'
+    property 'low'
+    property 'open'
+    property 'volume'
+  end
+end