+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.2.0 (Next)
 
+* [#1](https://github.com/dblock/google-finance-ruby-client/issues/1): Retrieve history of prices - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (2017/12/02)